// apps/bot/src/app/scenes/recipes/enums/recipes.enum.ts
export enum RECIPES_MAIN_BUTTON_CALLBACK {
  VEGAN_ICECREAM = 'RECIPES_VEGAN_ICECREAM',
  BROWNIE = 'RECIPES_BROWNIE',
  BOWL = 'RECIPES_BOWL',
  VEGAN_CHIA = 'RECIPES_VEGAN_CHIA',
  VEGAN_RATATOUILLE = 'RECIPES_VEGAN_RATATOUILLE',
  VEGAN_RISOTTO = 'RECIPES_VEGAN_RISOTTO',
  VEGAN_NAPOLEON = 'RECIPES_VEGAN_NAPOLEON',
  VEGAN_BREAD = 'RECIPES_VEGAN_BREAD',
  CABBAGE_TAHINI = 'RECIPES_CABBAGE_TAHINI',
}